refactor(playground-ui): move dynamic-form to lib/form
