+++ packages/playground-ui/src/domains/agents/components/AgentToolPanel.tsx
@@ -1,5 +1,5 @@
 import { usePlaygroundStore } from '@/store/playground-store';
-import { resolveSerializedZodOutput } from '@/components/dynamic-form/utils';
+import { resolveSerializedZodOutput } from '@/lib/form/utils';
 import { jsonSchemaToZod } from '@mastra/schema-compat/json-to-zod';
 import { parse } from 'superjson';
 import { z } from 'zod';
