+++ packages/playground-ui/src/lib/form/dynamic-form.tsx
@@ -1,11 +1,11 @@
 import { Loader2 } from 'lucide-react';
 import { Button } from '@/ds/components/Button';
-import { AutoForm } from '@/lib/form';
+import { AutoForm } from './auto-form';
 import type { ExtendableAutoFormProps } from '@autoform/react';
 import z, { ZodObject, ZodIntersection } from 'zod';
 import { Label } from '@/ds/components/Label';
 import { Icon } from '@/ds/icons';
-import { CustomZodProvider } from '@/lib/form/zod-provider';
+import { CustomZodProvider } from './zod-provider';
 
 interface DynamicFormProps<T extends z.ZodSchema> {
   schema: T;
