+++ packages/playground-ui/src/lib/form/index.ts
@@ -1,3 +1,4 @@
 export * from './auto-form';
+export * from './dynamic-form';
 export * from './types';
 export * from './utils';
